+++ src/Network/Network/ChangeLog.md
@@ -19,7 +19,7 @@
 --->
 
 ## Upcoming Release
-* Added support for `RecordTypes` property in `New-AzNetworkWatcherFlowLog` and `Set-AzNetworkWatcherFlowLog` commands
+* Added support for `RecordType` property in `New-AzNetworkWatcherFlowLog` and `Set-AzNetworkWatcherFlowLog` commands
 
 ## Version 7.22.0
 * Added new RouteTableUsageMode property for Network Manager Routing Configuration